moved Settings class to own file for better organization
