+++ espnff/__init__.py
@@ -1,4 +1,5 @@
 __all__ = ['League', 'Settings', 'Matchup', 'Team']
 
-from .espnff import League, Settings, Matchup
+from .espnff import League, Matchup
 from .team import Team
+from .settings import Settings
